Group test logic in separate files.
